src/flask-api/requirements.txt
# required python packages to run 
flask==2.0.2
python-dotenv==0.19.2
firebase-admin==5.2.0
kenpompy==0.3.0
# scikit-learn==1.0.1
pandas==1.3.5
pyrebase4==4.5.0
flask-cors==3.0.10
Flask-Session==0.4.0
# torch==1.10.2
gunicorn==20.1.0